chore: move schemas to separate folder
